Removed warnings
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@176280 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/dom/impl/DOMDocumentImpl.cpp
@@ -795,9 +795,9 @@ const XMLCh *  DOMDocumentImpl::getPooledString(const XMLCh *src)
 static const int kHeapAllocSize = 0x10000;    // The chunk size to allocate from the
                                               //   system allocator.
 
-static const int kMaxSubAllocationSize = 4096;   // Any request for more bytes
-                                                 //  than this will be handled by
-                                                 //  allocating directly with system.
+static const size_t kMaxSubAllocationSize = 4096;  // Any request for more bytes
+                                                   //  than this will be handled by
+                                                   //  allocating directly with system.
 
 void *         DOMDocumentImpl::allocate(size_t amount)
 {
